split input components and re-use for both dialogs
