+++ catalog/app/containers/Bucket/PackageCopyDialog.tsx
@@ -9,9 +9,7 @@ import * as NamedRoutes from 'utils/NamedRoutes'
 import StyledLink from 'utils/StyledLink'
 import * as workflows from 'utils/workflows'
 
-import SelectWorkflow from './PackageDialog/SelectWorkflow'
 import * as PD from './PackageDialog'
-import * as State from './PackageDialog/state'
 
 const useFormSkeletonStyles = M.makeStyles((t) => ({
   meta: {
@@ -69,126 +67,6 @@ const useStyles = M.makeStyles((t) => ({
   },
 }))
 
-function InputWorkflow() {
-  const {
-    formStatus,
-    metadataSchema: schema,
-    values: {
-      workflow: { status, value, onChange },
-    },
-    workflowsConfig,
-  } = State.use()
-  const error = React.useMemo(() => {
-    if (workflowsConfig._tag === 'error') return workflowsConfig.error.message
-    if (status._tag === 'error') return status.error.message
-    return undefined
-  }, [status, workflowsConfig])
-  if (workflowsConfig._tag === 'idle') return null
-  if (workflowsConfig._tag === 'loading') return <PD.WorkflowsInputSkeleton />
-  return (
-    <SelectWorkflow
-      disabled={schema._tag === 'loading' || formStatus._tag === 'submitting'}
-      error={error}
-      items={workflowsConfig.config.workflows}
-      onChange={onChange}
-      value={value}
-    />
-  )
-}
-
-function InputName() {
-  const {
-    formStatus,
-    values: {
-      name: { value, onChange, status },
-    },
-  } = State.use()
-  const handleChange = React.useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => onChange(event.target.value),
-    [onChange],
-  )
-  return (
-    <M.TextField
-      InputLabelProps={{ shrink: true }}
-      fullWidth
-      margin="normal"
-      helperText={<PD.PackageNameWarning />}
-      disabled={formStatus._tag === 'submitting'}
-      error={status._tag === 'error'}
-      label="Name"
-      placeholder="e.g. user/package"
-      onChange={handleChange}
-      value={value || ''}
-    />
-  )
-}
-
-function InputMessage() {
-  const {
-    formStatus,
-    values: {
-      message: { status, value, onChange },
-    },
-  } = State.use()
-  const handleChange = React.useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) => onChange(event.target.value),
-    [onChange],
-  )
-  return (
-    <M.TextField
-      InputLabelProps={{ shrink: true }}
-      fullWidth
-      margin="normal"
-      disabled={formStatus._tag === 'submitting'}
-      error={status._tag === 'error'}
-      helperText={status._tag === 'error' && status.error.message}
-      label="Message"
-      placeholder="Enter a commit message"
-      onChange={handleChange}
-      value={value || ''}
-    />
-  )
-}
-
-const useInputMetaStyles = M.makeStyles((t) => ({
-  root: {
-    display: 'flex',
-    flexDirection: 'column',
-    paddingTop: t.spacing(3),
-    overflowY: 'auto',
-  },
-}))
-
-const InputMeta = React.forwardRef<HTMLDivElement>(function InputMeta(_, ref) {
-  const classes = useInputMetaStyles()
-  const {
-    formStatus,
-    metadataSchema: schema,
-    values: {
-      meta: { status, value, onChange },
-    },
-  } = State.use()
-  const errors = React.useMemo(() => {
-    if (schema._tag === 'error') return [schema.error]
-    if (status._tag === 'error') return status.errors
-    return []
-  }, [schema, status])
-  if (schema._tag === 'loading') {
-    return <PD.MetaInputSkeleton ref={ref} className={classes.root} />
-  }
-  return (
-    <PD.MetaInput
-      disabled={formStatus._tag === 'submitting' || formStatus._tag === 'success'}
-      className={classes.root}
-      errors={errors}
-      onChange={onChange}
-      ref={ref}
-      schema={schema._tag === 'ready' ? schema.schema : undefined}
-      value={value}
-    />
-  )
-})
-
 interface FormErrorProps {
   error: Error
 }
@@ -210,7 +88,7 @@ interface PackageCopyFormProps {
 }
 
 function PackageCopyForm({ successor }: PackageCopyFormProps) {
-  const { params, formStatus, src, copy, progress } = State.use()
+  const { params, formStatus, src, copy, progress } = PD.useContext()
   const classes = useStyles()
   const [editorElement, setEditorElement] = React.useState<HTMLDivElement | null>(null)
   const { height: metaHeight = 0 } = useResizeObserver({ ref: editorElement })
@@ -220,10 +98,7 @@ function PackageCopyForm({ successor }: PackageCopyFormProps) {
     (event: React.FormEvent) => {
       event.preventDefault()
       invariant(src, 'Package handle must be provided')
-      invariant(
-        State.isPackageHandle(src),
-        'Full package handle with hash must be provided',
-      )
+      invariant(PD.isPackageHandle(src), 'Full package handle with hash must be provided')
       const destPrefix = successor.copyData && cfg.packageRoot ? cfg.packageRoot : null
       copy(src, destPrefix)
     },
@@ -235,10 +110,10 @@ function PackageCopyForm({ successor }: PackageCopyFormProps) {
       <DialogTitle bucket={successor.slug} />
       <M.DialogContent classes={dialogContentClasses}>
         <form className={classes.form} onSubmit={handleCopy}>
-          <InputWorkflow />
-          <InputName />
-          <InputMessage />
-          <InputMeta ref={setEditorElement} />
+          <PD.Inputs.Workflow />
+          <PD.Inputs.Name />
+          <PD.Inputs.Message />
+          <PD.Inputs.Meta ref={setEditorElement} />
           <input type="submit" style={{ display: 'none' }} />
         </form>
       </M.DialogContent>
@@ -328,7 +203,7 @@ export default function PackageCopyDialog({
   successor,
   onClose,
 }: PackageCopyDialogProps) {
-  const { formStatus, workflowsConfig, manifest, setOpen } = State.use()
+  const { formStatus, workflowsConfig, manifest, setOpen } = PD.useContext()
 
   React.useEffect(() => {
     setOpen(!!successor)
