+++ catalog/app/containers/Bucket/PackageDialog/Inputs/index.ts
@@ -1,0 +1,5 @@
+export { default as Name } from './Name'
+export { default as Message } from './Message'
+export { default as Workflow } from './Workflow'
+export { default as Meta } from './Meta'
+export { default as Files } from './Files'
