+++ catalog/app/containers/Bucket/PackageDialog/index.ts
@@ -2,13 +2,14 @@ export { default as DialogError } from './DialogError'
 export { default as DialogLoading } from './DialogLoading'
 export { default as DialogSuccess } from './DialogSuccess'
 export * from './FilesInput'
+export * as Inputs from './Inputs'
 export * from './Layout'
 export * from './Manifest'
 export * from './MetaInput'
 export * from './PackageCreationForm'
 export * from './PackageDialog'
 export * from './Skeleton'
-export { useContext, Provider } from './state'
+export { useContext, Provider, isPackageHandle } from './state'
 export { default as SubmitSpinner } from './SubmitSpinner'
 export * from './Uploads'
 export { default as useInitialActions } from './actions'
